see OPCUA-635

+++ FrameworkInternals/optionalModules.py
@@ -28,6 +28,8 @@ from shutil import copy, rmtree
 moduleInfo = {}
 
 
+#def stringR
+
 def getEnabledModules():
 	"""Get all enabled module metadata"""
 	baseDirectory = os.getcwd()
@@ -64,30 +66,35 @@ def listEnabledModules():
 	for module in enabledModules.keys():
 		print module, enabledModules[module]["tag"], "(requires quasar", enabledModules[module]["minVersion"]+")"
 
-def getModuleInfo():
-	"""Downloads list of modules from git and initializes global module list."""	
+def getModuleInfo(serverString="", forceFetch=False):
+	"""Downloads list of modules from git and initializes global module list."""
+
+	if not serverString: serverString = "https:://github.com"
+	else: forceFetch = True
 
 	baseDirectory = os.getcwd()
 	os.chdir(baseDirectory + os.path.sep + "FrameworkInternals")
 	#print("Changing directory to: " + baseDirectory + os.path.sep + "FrameworkInternals")
 	if not os.path.exists("quasar-modules"): os.mkdir("quasar-modules")
+	if forceFetch:
+		rmtree("quasar-modules")
+		os.mkdir("quasar-modules")
 	os.chdir("quasar-modules")
-	#print("Changing directory to: quasar-modules")
-	print("Checking out from git")
-	if os.path.exists(".git"):
+	print("Checking out module list from "+serverString)
+	if os.path.exists(".git") and not forceFetch:
 		try:
 			subprocess.call("git pull origin master" , shell=True)
 		except Exception, ex:
-			print "Error trying to fetch optional module list from github:", ex
+			print "Error trying to fetch optional module list from git:", ex
 			return False
 	else:
 		try:
 			subprocess.call("git init" , shell=True)
-			subprocess.call("git remote add origin https://github.com/quasar-team/quasar-modules.git" , shell=True)
+			subprocess.call("git remote add origin "+serverString+"/quasar-team/quasar-modules.git" , shell=True)
 			subprocess.call("git remote set-url --push origin push-disabled" , shell=True)
 			subprocess.call("git pull origin master" , shell=True)
 		except Exception, ex:
-			print "Error trying to fetch optional module list from github:", ex
+			print "Error trying to fetch optional module list from git:", ex
 			return False
 
 	moduleUrls = glob("*.url")
@@ -107,16 +114,18 @@ def getModuleInfo():
 	os.chdir(baseDirectory)
 	return True
 
-def enableModule(moduleName, tag="master"):
-	"""Enables optional module. Module URL and required quasar version is downloaded from github. Module download is done later at cmake configure stage. If tag argument exists, use it, otherwise use master head.
+def enableModule(moduleName, tag="master", serverString=""):
+	"""Enables optional module. Module URL and required quasar version is downloaded from github.
+	   Module download is done later at cmake configure stage.
 	
 	Keyword arguments:
-	moduleName -- name of the optional module
-	tag -- tag to checkout
+	moduleName   -- name of the optional module
+	tag          -- tag to checkout, if not specified, master branch is used
+	serverString -- default git server is "https://github.com", specify custom if necessary, e.g. "ssh://git@gitlab.cern.ch:7999"
 	"""	
 	print "Enabling module", moduleName, ", tag", tag
 
-	if not getModuleInfo(): return False
+	if not getModuleInfo(serverString): return False
 
 	print "Checking module to be compatible..."
 	quasarVersion = None
@@ -137,28 +146,61 @@ def enableModule(moduleName, tag="master"):
 	print("Copying module url file...")
 
 	baseDirectory = os.getcwd()
-	os.chdir(baseDirectory + os.path.sep + "FrameworkInternals")
+	fwInternalsDir = baseDirectory + os.path.sep + "FrameworkInternals"
+	os.chdir(fwInternalsDir)
+
+	# Check first if module is maybe already present with different version
+	#
+	tagFileName = "EnabledModules/"+moduleName+".tag"
+	oldTag = open(tagFileName).read()
+	if oldTag!=tag:
+		print "Old version of "+moduleName+" exists ("+oldTag+"). Removing it first..."
+		os.chdir(baseDirectory)
+		removeModule(moduleName)
+		os.chdir(fwInternalsDir)
+
+	# actually enable
+	#
 	if not os.path.isdir("EnabledModules"): os.mkdir("EnabledModules")
+	# FIXME: check if previous tag exists and possibly needs update
 	try:
 		for file in glob("quasar-modules/"+moduleName+".*"):
 			copy(file, "EnabledModules/")
+		# change URL if non-default server is specified
+		if serverString:
+			urlFileName = "EnabledModules/"+moduleName+".url"
+			url = open(urlFileName).read()
+			serverBegin = url.find("://")+3
+			serverEnd = url.find("/", serverBegin)
+			print "current server="+url[:serverEnd]
+			url = serverString+url[serverEnd:]
+			print "new url="+url
+			file = open(urlFileName, "w")
+			file.write(url)
 		# add tag
-		tagFileName = "EnabledModules/"+moduleName+".tag"
 		if os.path.exists(tagFileName): os.remove(tagFileName)
 		file = open(tagFileName, "w")
 		file.write(tag)
 	except Exception, ex:
 		print "Failed to set up module files in FrameworkInternals/EnabledModules/ :", ex
 		return False
+
 	os.chdir(baseDirectory)
 
 	print("Created module files.")
 
 	return True
 
+def replaceModuleGitServer(moduleName, serverString="github.com/"):
+	"""Changes the module source URL replacing the configured git server with the specified one.
+	
+	Keyword arguments:
+	moduleName -- name of the optional module
+	serverString -- string containing [user]"
+	"""	
 
 def disableModule(moduleName):
-	"""Disables optional module. Just disable the use of the module, no files will be deleted.
+	"""Disables optional module. Module files will be deleted.
 	
 	Keyword arguments:
 	moduleName -- name of the optional module
@@ -176,15 +218,37 @@ def disableModule(moduleName):
 
 	print("Removed url file of "+moduleName)
 	print("Remove module code if existing...")
-	cleanModule(moduleName)
+	removeModule(moduleName)
 
 	return True
 
-def cleanModule(module):
+def removeModule(module):
+	"""Removes optional module files without disabling it. Upon prepare_build or build the module will be freshly fetched.
+	
+	Keyword arguments:
+	moduleName -- name of the optional module
+	"""
+
+	# first check whether module contains modified files
+	#
+	baseDirectory = os.getcwd()
+	os.chdir(baseDirectory + os.path.sep + module)
+	output = subprocess.Popen(['git', 'ls-files', '-m'], stdout=subprocess.PIPE).communicate()[0]
+	## with python 2.7 use the following:
+	## output = subprocess.call("git ls-files -m", shell=True)
+	if output:
+		print "Error, tracked modified files exist in "+module+". Please resolve this before removing the module. Modified files: "
+		print output
+		return
+	os.chdir(baseDirectory)
+
+	# actually remove dirs and files
+	#
 	dirs = glob(module+"*")
+	dirs.extend(glob("FrameworkInternals/EnabledModules/"+module+"*/"))
 	if dirs:
 		print "Removing files of module", module
-		for dir in glob(module+"*"):
+		for dir in dirs:
 			print "Removing", dir
 			try:
 				rmtree(dir)
@@ -192,9 +256,9 @@ def cleanModule(module):
 				print "Failed to remove dir", dir, ex
 	else: print "Nothing to be removed for module", module
 
-def cleanModules():
+def removeModules():
 	"""Remove all enabled modules"""
 	enabledModules = getEnabledModules()
 	print "Removing downloaded modules"
 	for module in enabledModules.keys():
-		cleanModule(module)
+		removeModule(module)
